Move the ONNXModel class to a separate module
