+++ nncf/common/utils/backend.py
@@ -100,7 +100,7 @@ def is_onnx_model(model: Any) -> bool:
     """
     import onnx  # type: ignore
 
-    from nncf.onnx.graph.model_utils import ONNXModel  # type: ignore
+    from nncf.onnx.model import ONNXModel  # type: ignore
 
     return isinstance(model, onnx.ModelProto) or isinstance(model, ONNXModel)
